django_server/requirements.txt
pytesseract==0.3.10
opencv-python==4.6.0.66
Django==4.1.5
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
progressbar2==4.2.0
python-utils==3.4.5
requests==2.28.1
urllib3==1.26.13
Pillow==9.4.0
json-numpy==1.0.1
random-password-generator==2.2.0
torch>=1.7.0
torchvision>=0.8.1
pandas==2.0.3
ipython
psutil
PyYAML
tqdm
ultralytics
